Moved Cli code to an object called ConsoleCli which can be called from sbt console
